Github - Check Run ID
For some unknown reason, it seems that GitHub's check_run_id is losing its validity and cannot update the run interface.

With this, the Bamboo agent starts receiving a 500 when it validates check_run_id, in the previous version it might have received an HTTP CODE 200 and the interface did not update.

In this version, as it validated the existence of the check_run_id with GitHub, it was receiving an error "Octokit::UnprocessableEntity" - "422 - No commit found for SHA" and it ended up having to update the check_run to obtain a new ID.

Signed-off-by: Rodrigo Nardi <rnardi@netdef.org>