Old Checks support
In this PR we will deal with the old GitHub Checks.
When a PR runs and has the old entries, it will be marked as skipped and create the 3 new stages.
It also fixed a strange behavior when the build stage fails and leaves the testing stage in a peculiar state

Signed-off-by: Rodrigo Nardi <rnardi@netdef.org>